refactor search
