+++ src/index/btree/mod.rs
@@ -44,12 +44,14 @@ use parking_lot::RwLock;
 use std::any::Any;
 use std::marker::PhantomData;
 use index::btree::insert::*;
+use index::btree::search::*;
 
 mod cursor;
 mod external;
 mod internal;
 mod node;
 mod insert;
+mod search;
 
 const CACHE_SIZE: usize = 2048;
 
@@ -107,7 +109,7 @@ impl <KS, PS> BPlusTree<KS, PS>
     }
 
     pub fn seek(&self, key: &EntryKey, ordering: Ordering) -> RTCursor<KS, PS> {
-        let mut cursor = self.search(&self.get_root(), key, ordering);
+        let mut cursor = search_node(&self.get_root(), key, ordering);
         match ordering {
             Ordering::Forward => {}
             Ordering::Backward => {
@@ -124,43 +126,6 @@ impl <KS, PS> BPlusTree<KS, PS>
             }
         }
         cursor
-    }
-
-    fn search(&self, node_ref: &NodeCellRef, key: &EntryKey, ordering: Ordering) -> RTCursor<KS, PS> {
-        debug!("searching for {:?}", key);
-        read_node(node_ref, |node_handler: &NodeReadHandler<KS, PS>| {
-            let node = &**node_handler;
-            if let Some(right_node) = node.key_at_right_node(key) {
-                debug!("Search found a node at the right side");
-                return self.search(right_node, key, ordering);
-            }
-            let pos = node.search(key);
-            match node {
-                &NodeData::External(ref n) => {
-                    debug!(
-                        "search in external for {:?}, len {}, content: {:?}",
-                        key, n.len, n.keys
-                    );
-                    RTCursor::new(pos, node_ref, ordering)
-                },
-                &NodeData::Internal(ref n) => {
-                    debug!("search in internal node for {:?}, len {}", key, n.len);
-                    let next_node_ref = &n.ptrs.as_slice_immute()[pos];
-                    self.search(next_node_ref, key, ordering)
-                },
-                &NodeData::Empty(ref n) => {
-                    self.search(&n.right, key, ordering)
-                },
-                &NodeData::None => {
-                    RTCursor{
-                        index: 0,
-                        ordering,
-                        page: None,
-                        marker: PhantomData
-                    }
-                }
-            }
-        })
     }
 
     pub fn insert(&self, key: &EntryKey) {
